requirements.txt
pyaudio
oss2
requests
python-docx
lxml